requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django
django-allauth==0.40.0
django-braces==1.13.0
django-computedfields==0.0.10
django-cors-headers==3.1.1
django-crm==0.6.0
django-denorm==1.0.0
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-secure==1.0.1
django-sha2==0.4
django-stubs==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.0
djangorestframework-stubs==0.4.2
djoser==2.0.3
env-file==2019.4.13
idna==2.8
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
mypy==0.730
mypy-extensions==0.4.2
numpy==1.17.4
oauthlib==3.1.0
pbr==5.1.3
Pillow==6.2.0
public==2019.4.13
PyJWT==1.7.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2018.9
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.6
values==2019.4.13
vine==1.3.0
whitenoise==4.1.4
zipp==0.6.0
pandas
fpdf
matplotlib
django-sslserver
psycopg2